# openapi.yaml
openapi: 3.0.0
info:
  title: School Timetable Management API with Wellness
  version: 2.0.0
  description: |
    RESTful API for the School Timetable Management SaaS Platform, including automated timetable generation, 
    teacher wellness monitoring, workload management, and substitution handling. Supports multi-tenant schools 
    with role-based access (admin, principal, teacher, student, parent).
servers:
  - url: https://api.schooltimetable.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [email, role]
      properties:
        id:
          type: string
          format: uuid
        schoolId:
          type: string
          format: uuid
        email:
          type: string
        passwordHash:
          type: string
          writeOnly: true
        role:
          type: string
          enum: [admin, principal, teacher, student, parent]
        profile:
          type: object
        wellnessPreferences:
          type: object
        createdAt:
          type: string
          format: date-time
    Teacher:
      type: object
      required: [userId, subjects]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        subjects:
          type: array
          items:
            type: string
        availability:
          type: object
        preferences:
          type: object
        maxPeriodsPerDay:
          type: integer
          default: 6
        maxPeriodsPerWeek:
          type: integer
          default: 30
        maxConsecutivePeriods:
          type: integer
          default: 3
        minBreakDuration:
          type: integer
          default: 10
        wellnessScore:
          type: number
        burnoutRiskLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
...
    Class:
      type: object
      required: [name, grade]
      properties:
        id:
          type: string
          format: uuid
        schoolId:
          type: string
          format: uuid
        name:
          type: string
        grade:
          type: integer
        section:
          type: string
        stream:
          type: string
          enum: [Science, Commerce, Arts]
        studentCount:
          type: integer
...
    Timetable:
      type: object
      required: [schoolId, academicYearId]
      properties:
        id:
          type: string
          format: uuid
        schoolId:
          type: string
          format: uuid
        academicYearId:
          type: string
          format: uuid
        version:
          type: integer
        status:
          type: string
          enum: [draft, approved, active, archived]
        wellnessScore:
          type: number
        workloadBalanceScore:
          type: number
        metadata:
          type: object
        wellnessAnalysis:
          type: object
        createdBy:
          type: string
          format: uuid
        approvedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    TimetableWithWellness:
      type: object
      properties:
        timetable:
          $ref: '#/components/schemas/Timetable'
        wellnessAnalysis:
          type: object
          properties:
            overallScore:
              type: number
            teacherScores:
              type: object
            departmentScores:
              type: object
            issues:
              type: array
              items:
                type: object
            improvements:
              type: array
              items:
                type: string
...
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        details:
          type: object
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, role, schoolId]
              properties:
                email:
                  type: string
                password:
                  type: string
                  writeOnly: true
                role:
                  type: string
                  enum: [admin, principal, teacher, student, parent]
                schoolId:
                  type: string
                  format: uuid
                profile:
                  type: object
      responses:
        201:
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        401:
          $ref: '#/components/responses/Unauthorized'

  # Schools
...
  /users:
    get:
      summary: List users in school
      parameters:
        - name: schoolId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        201:
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    get:
      summary: Get user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Update user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        200:
          description: User updated
    delete:
      summary: Delete user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: User deleted

  # Teachers (extended users)
  /teachers:
    get:
      summary: List teachers
      parameters:
        - name: schoolId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
    post:
      summary: Create teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        201:
          description: Teacher created
...
  /classes:
    get:
      summary: List classes
      parameters:
        - name: schoolId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Class'
    post:
      summary: Create class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Class'
      responses:
        201:
          description: Class created
...
  /timetables/{id}/generate:
    post:
      summary: Generate timetable with wellness optimization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                options:
                  type: integer
                  default: 3
                timeout:
                  type: integer
                  default: 60
                optimization:
                  type: object
                  properties:
                    academicWeight:
                      type: number
                      default: 0.35
                    wellnessWeight:
                      type: number
                      default: 0.30
                    efficiencyWeight:
                      type: number
                      default: 0.20
                    preferenceWeight:
                      type: number
                      default: 0.15
      responses:
        200:
          description: Generated timetable options with wellness scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimetableWithWellness'
...